refactor: update filenames and enum based on PR feedback.
