+++ app/components/text-card.tsx
